# docs/api.yaml
openapi: 3.0.3
info:
  title: AidBridge API
  description: |
    AidBridge - Disaster Relief Coordinator API

    ## Overview
    RESTful API for coordinating disaster relief operations including:
    - SOS rescue requests
    - Aid supply requests and delivery
    - Volunteer mission management
    - Hub inventory management
    - Sponsor donations

    ## Authentication
    All protected endpoints require Bearer token in Authorization header.

    ## Database Version
    This API is aligned with database schema v4.0

    ## Caching Strategy

    The API uses Redis for caching frequently accessed data. Cache behavior is documented
    per-endpoint. General principles:

    ### Cache Patterns
    | Pattern | TTL | Use Case |
    |---------|-----|----------|
    | `mission:active:{id}` | 5m | Single active mission details |
    | `mission:active:list:{hash}` | 2m | Active missions list with filters |
    | `mission:stats:{scope}:{range}` | 5m | Aggregated statistics |
    | `user:{id}:active_missions` | 2m | User's active missions |

    ### Cache Invalidation
    - Write operations (POST/PUT/DELETE) invalidate relevant cache entries
    - Status changes trigger broadcast invalidation via Redis pub/sub
    - Clients should respect `cached_at` timestamps for cache-aware UX

    ### Endpoints NOT Cached
    - Audit trails (dispatch attempts, inventory logs)
    - Real-time data (tracking, chat)
    - Write operations
    - Paginated history queries

  version: 1.0.0
  contact:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token

  # ============================================================
  # ENUMS - Matching database schema v4.0 exactly
  # ============================================================
  schemas:
...
    HubStatus:
      type: string
      enum: [ACTIVE, INACTIVE, EMERGENCY]
      description: Hub operational status

    # ---------- ENUM: urgency_level ----------
    UrgencyLevel:
      type: string
      enum: [CRITICAL, HIGH, MEDIUM, LOW]
      description: Request urgency level

    # ---------- ENUM: sos_status ----------
...
    AidStatus:
      type: string
      enum: [PENDING, DISPATCHING, ASSIGNED, PICKED_UP, IN_TRANSIT, COMPLETED, CANCELLED]
      description: Aid request status

    # ---------- ENUM: donation_status ----------
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8
    Error:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input data
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
...
    CreateAidRequest:
      type: object
      required: [lat, lng, items]
      properties:
        sos_request_id:
          type: string
          format: uuid
          description: Link to SOS request if applicable
        lat:
          type: number
          format: double
          example: 10.762622
        lng:
          type: number
          format: double
          example: 106.660172
        address:
          type: string
        adults_count:
          type: integer
          minimum: 0
          default: 0
        elderly_count:
          type: integer
          minimum: 0
          default: 0
        children_count:
          type: integer
          minimum: 0
          default: 0
        notes:
          type: string
        urgency_level:
          $ref: '#/components/schemas/UrgencyLevel'
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AidRequestItemInput'
    AidRequestItemInput:
      type: object
      required: [item_category_id, quantity]
      properties:
        item_category_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
    AidRequestResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/AidRequest'
    AidRequest:
      type: object
      description: Aid request entity matching aid_requests table
      properties:
        id:
          type: string
          format: uuid
        requester_id:
          type: string
          format: uuid
        sos_request_id:
          type: string
          format: uuid
          nullable: true
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        address:
          type: string
          nullable: true
        adults_count:
          type: integer
        elderly_count:
          type: integer
        children_count:
          type: integer
        notes:
          type: string
          nullable: true
        urgency_level:
          $ref: '#/components/schemas/UrgencyLevel'
        status:
          $ref: '#/components/schemas/AidStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/AidRequestItem'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AidRequestItem:
      type: object
      description: Aid request line item matching aid_request_items table
      properties:
        id:
          type: string
          format: uuid
        aid_request_id:
          type: string
          format: uuid
        item_category_id:
          type: string
          format: uuid
        quantity:
          type: integer
        # Joined from item_categories
        category_name:
          type: string
        category_name_vi:
          type: string
        unit:
          type: string
        created_at:
          type: string
          format: date-time
    AidRequestListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/AidRequest'
            pagination:
              $ref: '#/components/schemas/Pagination'

    # ============================================================
    # MISSIONS - Matching missions + mission_items tables
    # ============================================================
...
    Hub:
      type: object
      description: Hub entity matching hubs table
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
        address:
          type: string
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        status:
          $ref: '#/components/schemas/HubStatus'
        contact_phone:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreateHubRequest:
      type: object
      required: [name, address, lat, lng]
      properties:
        name:
          type: string
          maxLength: 255
        address:
          type: string
          maxLength: 500
        phoneNumber:
          type: string
          maxLength: 50
        imageUrl:
          type: string
          maxLength: 500
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
          default: ACTIVE
        operatingHours:
          type: string
          maxLength: 255
        lat:
          type: number
          minimum: -90
          maximum: 90
        lng:
          type: number
          minimum: -180
          maximum: 180
    UpdateHubRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        address:
          type: string
          maxLength: 500
        phoneNumber:
          type: string
          maxLength: 50
        imageUrl:
          type: string
          maxLength: 500
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        operatingHours:
          type: string
          maxLength: 255
        lat:
          type: number
          minimum: -90
          maximum: 90
        lng:
          type: number
          minimum: -180
          maximum: 180
    HubResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Hub'
    HubListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/Hub'
    HubDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          allOf:
            - $ref: '#/components/schemas/Hub'
            - type: object
              properties:
                accepted_categories:
                  type: array
                  items:
                    $ref: '#/components/schemas/ItemCategory'
                inventory:
                  type: array
                  items:
                    $ref: '#/components/schemas/HubInventory'
                inventoryGroups:
                  type: array
                  items:
                    type: object
                    properties:
                      category:
                        type: string
                      parentCategoryName:
                        type: string
                      items:
                        type: array
                        items:
                          type: object
                          properties:
                            itemId:
                              type: string
                              format: uuid
                            name:
                              type: string
                            quantity:
                              type: integer
                            unit:
                              type: string
                            expiryDate:
                              type: string
                              nullable: true
                staff:
                  type: array
                  items:
                    $ref: '#/components/schemas/HubStaff'
    HubStaff:
      type: object
      description: Hub staff assignment matching hub_staff table
      properties:
        id:
          type: string
          format: uuid
        hub_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        is_available:
          type: boolean
        assigned_at:
          type: string
          format: date-time
        unassigned_at:
          type: string
          format: date-time
          nullable: true
        user:
          $ref: '#/components/schemas/UserProfile'

    # ============================================================
    # INVENTORY - Matching hub_inventories + inventory_logs
    # ============================================================
    HubInventory:
      type: object
      description: Hub inventory matching hub_inventories table
      properties:
        id:
          type: string
          format: uuid
        hub_id:
          type: string
          format: uuid
        item_category_id:
          type: string
          format: uuid
        current_quantity:
          type: integer
        low_stock_threshold:
          type: integer
        version:
          type: integer
          description: Optimistic locking version
        updated_at:
          type: string
          format: date-time
        category_name:
          type: string
        category_name_vi:
          type: string
        unit:
          type: string
        stock_level:
          type: string
          enum: [CRITICAL, LOW, NORMAL, HIGH]
...
    ItemCategory:
      type: object
      description: Item category matching item_categories table
      properties:
        id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        name_vi:
          type: string
        unit:
          type: string
        is_leaf:
          type: boolean
        created_at:
          type: string
          format: date-time
        children:
          type: array
          items:
            $ref: '#/components/schemas/ItemCategory'
    CategoryListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ItemCategory'
    CategoryDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          allOf:
            - $ref: '#/components/schemas/ItemCategory'
            - type: object
              properties:
                parent:
                  $ref: '#/components/schemas/ItemCategory'

    # ============================================================
    # CHAT - Matching chat_messages table
    # ============================================================
...
    AdminStaff:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        fullName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        hubId:
          type: string
          format: uuid
        hubName:
          type: string
    CreateAdminStaffRequest:
      type: object
      required: [fullName, email, phoneNumber, password, hubId]
      properties:
        fullName:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
        phoneNumber:
          type: string
          maxLength: 15
        password:
          type: string
          minLength: 8
          format: password
        hubId:
          type: string
          format: uuid
    AdminStaffListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminStaff'
    AdminStaffResponseWrapper:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/AdminStaff'

    # ============================================================
    # SYSTEM CONFIG - Matching system_config table
    # ============================================================
...
paths:
...
  /aid-requests:
    post:
      tags: [Aid Requests]
      summary: Create aid supply request
      operationId: createAidRequest
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAidRequest'
      responses:
        '201':
          description: Aid request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AidRequestResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags: [Aid Requests]
      summary: List aid requests
      operationId: listAidRequests
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AidStatus'
        - name: urgency_level
          in: query
          schema:
            $ref: '#/components/schemas/UrgencyLevel'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Aid requests list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AidRequestListResponse'
  /aid-requests/{id}:
    get:
      tags: [Aid Requests]
      summary: Get aid request details
      operationId: getAidRequest
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Aid request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AidRequestResponse'
        '404':
          description: Aid request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /aid-requests/{id}/cancel:
    post:
      tags: [Aid Requests]
      summary: Cancel aid request
      operationId: cancelAidRequest
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Aid request cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AidRequestResponse'
  /aid-requests/voice:
    post:
      tags: [Aid Requests]
      summary: Create aid request via voice (AI transcription)
      operationId: createVoiceAidRequest
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
                lat:
                  type: number
                lng:
                  type: number
      responses:
        '200':
          description: Voice transcribed and aid request suggested
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      transcription:
                        type: string
                      suggested_request:
                        $ref: '#/components/schemas/CreateAidRequest'
                      requires_confirmation:
                        type: boolean

  # ============================================================
  # MISSIONS
  # ============================================================
...
  /hubs:
    get:
      tags: [Hubs]
      summary: List hubs
      operationId: listHubs
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/HubStatus'
        - name: keyword
          in: query
          description: Search by hub name or address, case-insensitive
          schema:
            type: string
      responses:
        '200':
          description: Hubs list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HubListResponse'

    post:
      tags: [Hubs]
      summary: Create new hub (Admin only)
      operationId: createHub
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHubRequest'
      responses:
        '201':
          description: Hub created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HubResponse'
  /hubs/{id}:
    get:
      tags: [Hubs]
      summary: Get hub details
      operationId: getHub
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Hub details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HubDetailResponse'

    patch:
      tags: [Hubs]
      summary: Update hub (Admin only)
      operationId: updateHub
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateHubRequest'
      responses:
        '200':
          description: Hub updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HubResponse'
...
  /categories:
    get:
      tags: [Categories]
      summary: Get item categories tree
      operationId: listCategories
      parameters:
        - name: is_leaf
          in: query
          schema:
            type: boolean
          description: Filter leaf categories only
      responses:
        '200':
          description: Categories tree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
  /categories/{id}:
    get:
      tags: [Categories]
      summary: Get category details
      operationId: getCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Category details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDetailResponse'

  # ============================================================
  # NOTIFICATIONS
  # ============================================================
...
  /admin/staff:
    get:
      tags: [Admin]
      summary: List staff accounts with assigned hub
      operationId: listAdminStaff
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Staff list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStaffListResponse'

    post:
      tags: [Admin]
      summary: Create a staff account and assign it to a hub
      operationId: createAdminStaff
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdminStaffRequest'
      responses:
        '201':
          description: Staff created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStaffResponseWrapper'
...